dlog: resource limits pt6: default config is the dynamic fallback
RB_ID=777357

+++ distributedlog-core/src/main/java/com/twitter/distributedlog/rate/MovingAverageRate.java
@@ -2,6 +2,6 @@ package com.twitter.distributedlog.rate;
 
 public interface MovingAverageRate {
     double get();
-    void add(int amount);
+    void add(long amount);
     void inc();
 }
