+++ distributedlog-core/src/main/java/com/twitter/distributedlog/rate/SampledMovingAverageRate.java
@@ -4,16 +4,16 @@ import com.twitter.common.stats.Rate;
 import com.twitter.util.TimerTask;
 import com.twitter.util.Timer;
 import com.twitter.util.Time;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 class SampledMovingAverageRate implements MovingAverageRate {
     private final Rate rate;
-    private final AtomicInteger total;
+    private final AtomicLong total;
 
     private double value;
 
     public SampledMovingAverageRate(int intervalSecs) {
-        this.total = new AtomicInteger(0);
+        this.total = new AtomicLong(0);
         this.rate = Rate.of("Ignore", total)
             .withWindowSize(intervalSecs)
             .build();
@@ -26,7 +26,7 @@ class SampledMovingAverageRate implements MovingAverageRate {
     }
 
     @Override
-    public void add(int amount) {
+    public void add(long amount) {
         total.getAndAdd(amount);
     }
